# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RAHATI Platform API
  version: "1.0.0-phase1-draft"
  description: >
    RAHATI digital platform REST API (ADR-0007). Contract-first per API-First principle —
    authored before controller implementation (Phase 4). Every operation cites the SRS
    FR-*/NFR-* requirement(s) it implements. See docs/api/api-architecture.md for conventions.
  contact:
    name: RAHATI Engineering
servers:
  - url: https://api.raahti.dz/v1
    description: Production (indicative — hosting provider pending ADR-0016)
  - url: https://staging-api.raahti.dz/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /stations/{stationId}:
    get:
      tags: [Places]
      summary: Get station detail (place detail sheet, RAHETI unit)
      description: "Implements FR-PLC-01..05 for a Station. IoT-sourced real-time status (FR-PLC-02)."
      security: []
      parameters:
        - name: stationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Station detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StationDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /stations/{stationId}/cabins:
    get:
      tags: [Places]
      summary: List cabins for a station with real-time occupancy
      description: "Implements FR-PLC-02, feeds FR-PAY-01/02 QR flow entry."
      security: []
      parameters:
        - name: stationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Cabin list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Cabin" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth-issued JWT with role/site_scope custom claims (ADR-0009).
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    TooManyRequests:
      description: >
        Rate limit exceeded (api-architecture.md §9). Per-caller (authenticated user, or
        IP for guest/public endpoints), per-route token bucket. `Retry-After` gives the
        number of seconds until at least one request is allowed again.
      headers:
        Retry-After:
          description: Seconds until the caller may retry.
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
...
  schemas:
    ProblemDetail:
      type: object
      description: RFC 7807 — see api-architecture.md §7.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
    BilingualText:
      type: object
      description: >
        Trilingual (FR/AR/EN) as of ADR-0017 (Phase 2) / Phase 3 Feature 1 —
        name retained for historical continuity with the Phase 1 draft; all
        three fields are required on every response.
      properties:
        fr: { type: string }
        ar: { type: string }
        en: { type: string }
      required: [fr, ar, en]
    Money:
      type: object
      properties:
        amount: { type: string, description: "Decimal as string — avoids float rounding." }
        currency: { type: string, example: "DZD" }
      required: [amount, currency]
    GeoPoint:
      type: object
      properties:
        type: { type: string, enum: [Point] }
        coordinates:
          type: array
          items: { type: number }
          minItems: 2
          maxItems: 2
          description: "[lng, lat] — GeoJSON/PostGIS order."
...
    PlaceSummary:
      type: object
      description: Unified Station + Third-Party Place map/list read model (System Architecture §7).
      properties:
        id: { type: string, format: uuid }
        placeKind: { type: string, enum: [station, third_party_place] }
        name: { $ref: "#/components/schemas/BilingualText" }
        position: { $ref: "#/components/schemas/GeoPoint" }
        pinColor: { type: string, enum: [green, blue, amber, magenta], description: "RAH-DOC-002 §4.2 functional color coding." }
        distanceMeters: { type: number }
        averageRating: { type: number, nullable: true }
        reviewCount: { type: integer }
        isFree: { type: boolean }
        tags: { type: array, items: { type: string } }
    StationDetail:
      allOf:
        - $ref: "#/components/schemas/PlaceSummary"
        - type: object
          properties:
            configuration: { type: string, enum: [fixe, mobile, event] }
            status: { type: string, enum: [active, inactive, maintenance] }
            cabins: { type: array, items: { $ref: "#/components/schemas/Cabin" } }
            slatokiTent: { $ref: "#/components/schemas/SlatokiTent" }
...
    Cabin:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        type: { type: string, enum: [H, F, Slatoki, PMR] }
        occupancyStatus: { type: string, enum: [free, occupied, out_of_service] }
        isPaid: { type: boolean }
        price: { $ref: "#/components/schemas/Money" }
    SlatokiTent:
      type: object
      nullable: true
      properties:
        deploymentStatus: { type: string, enum: [deployed, folded] }
        matCapacity: { type: integer }
        hasLighting: { type: boolean }
        hasPrivacyCurtain: { type: boolean }
...